Split up OAuth2ClientAutoConfigurationTests
