+++ spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/security/oauth2/client/ClientRegistrationRepositoryAutoConfigurationTests.java
@@ -16,19 +16,14 @@
 
 package org.springframework.boot.autoconfigure.security.oauth2.client;
 
-import java.lang.reflect.Field;
-import java.net.URI;
-import java.util.List;
-import java.util.Map;
+import static org.assertj.core.api.Assertions.assertThat;
+
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-
-import javax.servlet.Filter;
 
 import org.junit.After;
 import org.junit.Test;
 import org.mockito.Mockito;
-
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.NoSuchBeanDefinitionException;
 import org.springframework.boot.SpringApplication;
@@ -40,32 +35,18 @@ import org.springframework.boot.test.util.TestPropertyValues;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.ProviderManager;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.oauth2.client.authentication.AuthorizationCodeAuthenticationProcessingFilter;
-import org.springframework.security.oauth2.client.authentication.AuthorizationCodeAuthenticationProvider;
-import org.springframework.security.oauth2.client.authentication.AuthorizationCodeRequestRedirectFilter;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
-import org.springframework.security.oauth2.client.user.nimbus.NimbusOAuth2UserService;
 import org.springframework.security.oauth2.core.AuthorizationGrantType;
 import org.springframework.security.oauth2.core.ClientAuthenticationMethod;
-import org.springframework.security.web.DefaultSecurityFilterChain;
-import org.springframework.security.web.FilterChainProxy;
-import org.springframework.security.web.util.matcher.AnyRequestMatcher;
-import org.springframework.util.ReflectionUtils;
 import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 /**
- * Tests for {@link OAuth2ClientAutoConfiguration}.
+ * Tests for {@link ClientRegistrationRepositoryAutoConfiguration}.
  *
  * @author Joe Grandja
  */
-public class OAuth2ClientAutoConfigurationTests {
+public class ClientRegistrationRepositoryAutoConfigurationTests {
 	private static final String CLIENT_PROPERTY_PREFIX = "security.oauth2.client";
 
 	private static final String CLIENT_ID_PROPERTY = "client-id";
@@ -151,7 +132,6 @@ public class OAuth2ClientAutoConfigurationTests {
 		this.processEnvironment();
 		this.context.refresh();
 
-		// ***** Assert -> ClientRegistrationRepositoryConfiguration
 		ClientRegistrationRepository clientRegistrationRepository = this
 				.getBean(ClientRegistrationRepository.class);
 		assertThat(clientRegistrationRepository).isNotNull();
@@ -197,45 +177,6 @@ public class OAuth2ClientAutoConfigurationTests {
 		assertThat(facebookClientRegistration.getClientSecret())
 				.isEqualTo("facebook-client-secret");
 		this.assertFacebookClientPropertyDefaults(facebookClientRegistration);
-
-		// ***** Assert -> OAuth2LoginConfiguration
-		List<Filter> oauth2SecurityFilters = this.assertOAuth2SecurityFiltersConfigured();
-
-		AuthorizationCodeRequestRedirectFilter authorizationCodeRequestRedirectFilter = (AuthorizationCodeRequestRedirectFilter) oauth2SecurityFilters
-				.get(0);
-		AuthorizationCodeAuthenticationProcessingFilter authorizationCodeAuthenticationProcessingFilter = (AuthorizationCodeAuthenticationProcessingFilter) oauth2SecurityFilters
-				.get(1);
-
-		assertThat(this.reflectClientRegistrationRepository(
-				authorizationCodeRequestRedirectFilter))
-						.isSameAs(clientRegistrationRepository);
-		assertThat(this.reflectClientRegistrationRepository(
-				authorizationCodeAuthenticationProcessingFilter))
-						.isSameAs(clientRegistrationRepository);
-
-		AuthorizationCodeAuthenticationProvider authorizationCodeAuthenticationProvider = this
-				.reflectAuthenticationProvider(
-						authorizationCodeAuthenticationProcessingFilter);
-		assertThat(authorizationCodeAuthenticationProvider).isNotNull();
-
-		NimbusOAuth2UserService nimbusOAuth2UserService = this
-				.reflectOAuth2UserService(authorizationCodeAuthenticationProvider);
-		assertThat(nimbusOAuth2UserService).isNotNull();
-
-		Field userNameAttributeNamesField = ReflectionUtils
-				.findField(NimbusOAuth2UserService.class, "userNameAttributeNames");
-		ReflectionUtils.makeAccessible(userNameAttributeNamesField);
-		Map<URI, String> userNameAttributeNames = (Map<URI, String>) ReflectionUtils
-				.getField(userNameAttributeNamesField, nimbusOAuth2UserService);
-		assertThat(userNameAttributeNames).isNotEmpty();
-
-		URI gitHubUserInfoUri = URI.create("https://api.github.com/user");
-		assertThat(userNameAttributeNames.containsKey(gitHubUserInfoUri)).isTrue();
-		assertThat(userNameAttributeNames.get(gitHubUserInfoUri)).isEqualTo("name");
-
-		URI facebookUserInfoUri = URI.create("https://graph.facebook.com/me");
-		assertThat(userNameAttributeNames.containsKey(facebookUserInfoUri)).isTrue();
-		assertThat(userNameAttributeNames.get(facebookUserInfoUri)).isEqualTo("name");
 	}
 
 	@Test
@@ -254,7 +195,6 @@ public class OAuth2ClientAutoConfigurationTests {
 		this.processEnvironment();
 		this.context.refresh();
 
-		// ***** Assert -> ClientRegistrationRepositoryConfiguration
 		ClientRegistrationRepository clientRegistrationRepository = this
 				.getBean(ClientRegistrationRepository.class);
 		assertThat(clientRegistrationRepository).isNotNull();
@@ -267,32 +207,17 @@ public class OAuth2ClientAutoConfigurationTests {
 		assertThat(googleClientRegistration.getClientSecret())
 				.isEqualTo("google-client-secret");
 		this.assertGoogleClientPropertyDefaults(googleClientRegistration);
-
-		// ***** Assert -> OAuth2LoginConfiguration
-		List<Filter> oauth2SecurityFilters = this.assertOAuth2SecurityFiltersConfigured();
-
-		AuthorizationCodeRequestRedirectFilter authorizationCodeRequestRedirectFilter = (AuthorizationCodeRequestRedirectFilter) oauth2SecurityFilters
-				.get(0);
-		AuthorizationCodeAuthenticationProcessingFilter authorizationCodeAuthenticationProcessingFilter = (AuthorizationCodeAuthenticationProcessingFilter) oauth2SecurityFilters
-				.get(1);
-
-		assertThat(this.reflectClientRegistrationRepository(
-				authorizationCodeRequestRedirectFilter))
-						.isSameAs(clientRegistrationRepository);
-		assertThat(this.reflectClientRegistrationRepository(
-				authorizationCodeAuthenticationProcessingFilter))
-						.isSameAs(clientRegistrationRepository);
 	}
 
 	@Test(expected = NoSuchBeanDefinitionException.class)
 	public void refreshContextWhenNotWebContextThenConfigurationBacksOff()
 			throws Exception {
 		AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
-		context.register(OAuth2ClientAutoConfiguration.class,
+		context.register(ClientRegistrationRepositoryAutoConfiguration.class,
 				SecurityAutoConfiguration.class,
 				PropertyPlaceholderAutoConfiguration.class);
 		context.refresh();
-		context.getBean(FilterChainProxy.class);
+		context.getBean(ClientRegistrationRepository.class);
 	}
 
 	@Test
@@ -301,7 +226,9 @@ public class OAuth2ClientAutoConfigurationTests {
 		this.prepareContext(DefaultConfiguration.class);
 		this.processEnvironment();
 		this.context.refresh();
-		this.assertDefaultConfigurationBacksOff();
+		ClientRegistrationRepository clientRegistrationRepository = this
+				.getBean(ClientRegistrationRepository.class);
+		assertThat(clientRegistrationRepository).isNull();
 	}
 
 	@Test
@@ -317,7 +244,9 @@ public class OAuth2ClientAutoConfigurationTests {
 
 		this.processEnvironment();
 		this.context.refresh();
-		this.assertDefaultConfigurationBacksOff();
+		ClientRegistrationRepository clientRegistrationRepository = this
+				.getBean(ClientRegistrationRepository.class);
+		assertThat(clientRegistrationRepository).isNull();
 	}
 
 	@Test
@@ -337,11 +266,13 @@ public class OAuth2ClientAutoConfigurationTests {
 
 		this.processEnvironment();
 		this.context.refresh();
-		this.assertDefaultConfigurationBacksOff();
+		ClientRegistrationRepository clientRegistrationRepository = this
+				.getBean(ClientRegistrationRepository.class);
+		assertThat(clientRegistrationRepository).isNull();
 	}
 
 	@Test
-	public void refreshContextWhenClientRegistrationRepositoryOverrideThenClientRegistrationRepositoryConfigurationBacksOff()
+	public void refreshContextWhenClientRegistrationRepositoryOverrideThenConfigurationBacksOff()
 			throws Exception {
 		ClientRegistration clientRegistration = Mockito.mock(ClientRegistration.class);
 		Mockito.when(clientRegistration.getClientAlias()).thenReturn("mock-client-alias");
@@ -353,9 +284,9 @@ public class OAuth2ClientAutoConfigurationTests {
 				ClientRegistrationRepository.class);
 		Mockito.when(expectedClientRegistrationRepository.getRegistrations())
 				.thenReturn(Stream.of(clientRegistration).collect(Collectors.toList()));
-		CustomClientRegistrationRepositoryConfiguration.clientRegistrationRepository = expectedClientRegistrationRepository;
+		CustomConfiguration.clientRegistrationRepository = expectedClientRegistrationRepository;
 
-		this.prepareContext(CustomClientRegistrationRepositoryConfiguration.class);
+		this.prepareContext(CustomConfiguration.class);
 
 		// Prepare environment
 		TestPropertyValues
@@ -373,29 +304,6 @@ public class OAuth2ClientAutoConfigurationTests {
 		assertThat(clientRegistrationRepository).isNotNull();
 		assertThat(clientRegistrationRepository)
 				.isSameAs(expectedClientRegistrationRepository);
-	}
-
-	@Test
-	public void refreshContextWhenOAuth2LoginOverrideThenOAuth2LoginConfigurationBacksOff()
-			throws Exception {
-		this.prepareContext(CustomOAuth2LoginConfiguration.class);
-
-		// Prepare environment
-		TestPropertyValues
-				.of(GOOGLE_CLIENT_PROPERTY_BASE + "." + CLIENT_ID_PROPERTY
-						+ "=google-client-id")
-				.and(GOOGLE_CLIENT_PROPERTY_BASE + "." + CLIENT_SECRET_PROPERTY
-						+ "=google-client-secret")
-				.applyTo(this.context.getEnvironment());
-
-		this.processEnvironment();
-		this.context.refresh();
-
-		this.assertOAuth2SecurityFiltersConfigured();
-		CustomOAuth2LoginConfiguration customOAuth2LoginConfig = this
-				.getBean(CustomOAuth2LoginConfiguration.class);
-		assertThat(customOAuth2LoginConfig).isNotNull();
-		assertThat(customOAuth2LoginConfig.configured).isTrue();
 	}
 
 	private void assertGoogleClientPropertyDefaults(
@@ -475,106 +383,13 @@ public class OAuth2ClientAutoConfigurationTests {
 		assertThat(clientRegistration.getClientAlias()).isEqualTo(FACEBOOK_CLIENT_ALIAS);
 	}
 
-	private List<Filter> assertOAuth2SecurityFiltersConfigured() {
-		FilterChainProxy springSecurityFilterChain = this.getBean(FilterChainProxy.class);
-		assertThat(springSecurityFilterChain).isNotNull();
-		assertThat(springSecurityFilterChain.getFilterChains().size()).isEqualTo(1);
-
-		DefaultSecurityFilterChain securityFilterChain = (DefaultSecurityFilterChain) springSecurityFilterChain
-				.getFilterChains().get(0);
-		assertThat(securityFilterChain.getRequestMatcher())
-				.isSameAs(AnyRequestMatcher.INSTANCE);
-		List<Filter> oauth2SecurityFilters = securityFilterChain.getFilters().stream()
-				.filter(e -> e.getClass()
-						.equals(AuthorizationCodeRequestRedirectFilter.class)
-						|| e.getClass().equals(
-								AuthorizationCodeAuthenticationProcessingFilter.class))
-				.collect(Collectors.toList());
-		assertThat(oauth2SecurityFilters).isNotEmpty();
-		assertThat(oauth2SecurityFilters.size()).isEqualTo(2);
-		assertThat(oauth2SecurityFilters.get(0))
-				.isInstanceOf(AuthorizationCodeRequestRedirectFilter.class);
-		assertThat(oauth2SecurityFilters.get(1))
-				.isInstanceOf(AuthorizationCodeAuthenticationProcessingFilter.class);
-
-		return oauth2SecurityFilters;
-	}
-
-	private void assertDefaultConfigurationBacksOff() {
-		ClientRegistrationRepository clientRegistrationRepository = this
-				.getBean(ClientRegistrationRepository.class);
-		assertThat(clientRegistrationRepository).isNull();
-
-		FilterChainProxy springSecurityFilterChain = this.getBean(FilterChainProxy.class);
-		assertThat(springSecurityFilterChain).isNotNull(); // Boot security configuration
-															// in effect
-
-		springSecurityFilterChain.getFilterChains().forEach(chain -> {
-			List<Filter> oauth2SecurityFilters = chain.getFilters().stream().filter(e -> e
-					.getClass().equals(AuthorizationCodeRequestRedirectFilter.class)
-					|| e.getClass().equals(
-							AuthorizationCodeAuthenticationProcessingFilter.class))
-					.collect(Collectors.toList());
-			assertThat(oauth2SecurityFilters).isEmpty();
-		});
-	}
-
-	private ClientRegistrationRepository reflectClientRegistrationRepository(
-			AuthorizationCodeRequestRedirectFilter authorizationRedirectFilter) {
-
-		Field clientRegistrationRepositoryField = ReflectionUtils.findField(
-				AuthorizationCodeRequestRedirectFilter.class,
-				"clientRegistrationRepository");
-		ReflectionUtils.makeAccessible(clientRegistrationRepositoryField);
-		return (ClientRegistrationRepository) ReflectionUtils
-				.getField(clientRegistrationRepositoryField, authorizationRedirectFilter);
-	}
-
-	private ClientRegistrationRepository reflectClientRegistrationRepository(
-			AuthorizationCodeAuthenticationProcessingFilter authenticationProcessingFilter) {
-
-		Field clientRegistrationRepositoryField = ReflectionUtils.findField(
-				AuthorizationCodeAuthenticationProcessingFilter.class,
-				"clientRegistrationRepository");
-		ReflectionUtils.makeAccessible(clientRegistrationRepositoryField);
-		return (ClientRegistrationRepository) ReflectionUtils.getField(
-				clientRegistrationRepositoryField, authenticationProcessingFilter);
-	}
-
-	private AuthorizationCodeAuthenticationProvider reflectAuthenticationProvider(
-			AuthorizationCodeAuthenticationProcessingFilter authenticationProcessingFilter) {
-
-		Field authenticationMangerField = ReflectionUtils.findField(
-				AuthorizationCodeAuthenticationProcessingFilter.class.getSuperclass(),
-				"authenticationManager");
-		ReflectionUtils.makeAccessible(authenticationMangerField);
-		ProviderManager authenticationManager = (ProviderManager) ReflectionUtils
-				.getField(authenticationMangerField, authenticationProcessingFilter);
-
-		return (AuthorizationCodeAuthenticationProvider) authenticationManager
-				.getProviders().stream()
-				.filter(e -> e.getClass()
-						.equals(AuthorizationCodeAuthenticationProvider.class))
-				.findFirst().orElseGet(null);
-	}
-
-	private NimbusOAuth2UserService reflectOAuth2UserService(
-			AuthorizationCodeAuthenticationProvider authenticationProvider) {
-		Field userInfoServiceField = ReflectionUtils.findField(
-				AuthorizationCodeAuthenticationProvider.class, "userInfoService");
-		ReflectionUtils.makeAccessible(userInfoServiceField);
-		return (NimbusOAuth2UserService) ReflectionUtils.getField(userInfoServiceField,
-				authenticationProvider);
-	}
-
 	private void prepareContext(Class<?>... configurationClasses) {
 		this.context = new AnnotationConfigWebApplicationContext();
 		if (configurationClasses != null) {
 			this.context.register(configurationClasses);
 		}
-		this.context.register(OAuth2ClientAutoConfiguration.class,
-				SecurityAutoConfiguration.class,
-				PropertyPlaceholderAutoConfiguration.class);
+		this.context.register(ClientRegistrationRepositoryAutoConfiguration.class,
+				SecurityAutoConfiguration.class, PropertyPlaceholderAutoConfiguration.class);
 	}
 
 	private void processEnvironment() {
@@ -601,30 +416,12 @@ public class OAuth2ClientAutoConfigurationTests {
 	}
 
 	@Configuration
-	protected static class CustomClientRegistrationRepositoryConfiguration {
+	protected static class CustomConfiguration {
 		private static ClientRegistrationRepository clientRegistrationRepository;
 
 		@Bean
 		protected ClientRegistrationRepository clientRegistrationRepository() {
 			return clientRegistrationRepository;
 		}
-	}
-
-	@EnableWebSecurity
-	protected static class CustomOAuth2LoginConfiguration
-			extends WebSecurityConfigurerAdapter {
-		private boolean configured;
-
-		// @formatter:off
-		@Override
-		protected void configure(HttpSecurity http) throws Exception {
-			http
-				.authorizeRequests()
-					.anyRequest().authenticated()
-					.and()
-				.oauth2Login();
-			this.configured = true;
-		}
-		// @formatter:on
 	}
 }
